fixes done
