--- backend/src/password/PasswordEncryptr.ts
@@ -1,28 +1,0 @@
-import { User } from '@entity/User'
-// import { logger } from '@test/testSetup' getting error "jest is not defined"
-import { getBasicCryptographicKey, SecretKeyCryptographyCreateKey } from './EncryptorUtils'
-
-export const encryptPassword = (dbUser: User, password: string): bigint => {
-  const basicKey = getBasicCryptographicKey(dbUser)
-  if (!basicKey) {
-    // logger.error('Password not set for user ' + dbUser.id)
-    throw new Error('Password not set for user ' + dbUser.id) // user has no password
-  } else {
-    const keyBuffer = SecretKeyCryptographyCreateKey(basicKey, password) // return short and long hash
-    const passwordHash = keyBuffer[0].readBigUInt64LE()
-    return passwordHash
-  }
-}
-
-export const verifyPassword = (dbUser: User, password: string): boolean => {
-  const basicKey = getBasicCryptographicKey(dbUser)
-  if (!basicKey) {
-    // logger.error('Password not set for user ' + dbUser.id)
-    throw new Error('Password not set for user ' + dbUser.id) // user has no password
-  } else {
-    if (dbUser.password.toString() !== encryptPassword(dbUser, password).toString()) {
-      return false
-    }
-    return true
-  }
-}
